Fix: ShopScrollResponse.totalScore 데이터 타입 변경 Double -> BigDecimal

+++ src/main/java/toy/yogiyo/core/shop/dto/ShopScrollResponse.java
@@ -3,6 +3,8 @@ package toy.yogiyo.core.shop.dto;
 import lombok.*;
 import toy.yogiyo.core.shop.domain.Shop;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor(access = AccessLevel.PUBLIC)
 @AllArgsConstructor
@@ -11,7 +13,7 @@ public class ShopScrollResponse {
 
     private Long shopId;
     private String shopName;
-    private Double totalScore;
+    private BigDecimal totalScore;
     private Double distance;
     private Integer minDeliveryTime;
     private Integer maxDeliveryTime;
